Separate component for selecting values
